chore: Renamed components
