+++ docs/embroider-css-modules/app/templates/dashboard.hbs
@@ -1,6 +1,6 @@
 {{page-title "Dashboard"}}
 
-<Ui::PageNew @title="Dashboard">
+<Ui::Page @title="Dashboard">
   <div class={{this.styles.widgets}}>
     <div class={{this.styles.widget-1}} data-test-widget="1">
       <Widgets::Widget-1 />
@@ -22,4 +22,4 @@
       <Widgets::Widget-5 />
     </div>
   </div>
-</Ui::PageNew>
+</Ui::Page>
