+++ docs/embroider-css-modules/app/templates/form.hbs
@@ -1,6 +1,6 @@
 {{page-title "Form"}}
 
-<Ui::PageNew @title="Form">
+<Ui::Page @title="Form">
   <Ui::Form
     @data={{hash
       email=undefined
@@ -42,4 +42,4 @@
       />
     </div>
   </Ui::Form>
-</Ui::PageNew>
+</Ui::Page>
