+++ docs/embroider-css-modules/app/templates/index.hbs
@@ -1,4 +1,4 @@
-<Ui::PageNew @title="Welcome!">
+<Ui::Page @title="Welcome!">
   <p>
     Thanks for trying out
     <a
@@ -16,4 +16,4 @@
     To see what you can do with CSS modules, visit one of the
     examples and inspect the DOM.
   </p>
-</Ui::PageNew>
+</Ui::Page>
