+++ docs/embroider-css-modules/app/templates/not-found.hbs
@@ -1,6 +1,6 @@
 {{page-title "Page Not Found"}}
 
-<Ui::PageNew @title="404">
+<Ui::Page @title="404">
   <p>Feeling lost? Un<em>contained</em>?</p>
   <p>Don't worry. We all have our off days.</p>
 
@@ -28,4 +28,4 @@
       </div>
     </ContainerQuery>
   </div>
-</Ui::PageNew>
+</Ui::Page>
