+++ docs/embroider-css-modules/app/templates/products.hbs
@@ -1,6 +1,6 @@
 {{page-title "Products"}}
 
-<Ui::PageNew @title="Products">
+<Ui::Page @title="Products">
   <div
     class={{local-class
       this.styles
@@ -38,4 +38,4 @@
       {{outlet}}
     </div>
   </div>
-</Ui::PageNew>
+</Ui::Page>
