+++ docs/sample-v2-addon/package.json
@@ -107,8 +107,8 @@
   },
   "ember-addon": {
     "app-js": {
-      "./components/navigation-menu-new.js": "./dist/_app_/components/navigation-menu-new.js",
-      "./components/ui/page-new.js": "./dist/_app_/components/ui/page-new.js"
+      "./components/navigation-menu.js": "./dist/_app_/components/navigation-menu.js",
+      "./components/ui/page.js": "./dist/_app_/components/ui/page.js"
     },
     "main": "addon-main.cjs",
     "type": "addon",
