--- docs/sample-v2-addon/src/components/ui/page-new.ts
@@ -1,16 +1,0 @@
-import Component from '@glimmer/component';
-
-import styles from './page-new.css';
-
-interface UiPageNewSignature {
-  Args: {
-    title: string;
-  };
-  Blocks: {
-    default: [];
-  };
-}
-
-export default class UiPageNewComponent extends Component<UiPageNewSignature> {
-  styles = styles;
-}
