+++ docs/sample-v2-addon/src/components/ui/page.css.d.ts
