+++ docs/sample-v2-addon/src/components/ui/page.hbs
