+++ docs/sample-v2-addon/src/components/ui/page.ts
@@ -1,0 +1,16 @@
+import Component from '@glimmer/component';
+
+import styles from './page.css';
+
+interface UiPageSignature {
+  Args: {
+    title: string;
+  };
+  Blocks: {
+    default: [];
+  };
+}
+
+export default class UiPageComponent extends Component<UiPageSignature> {
+  styles = styles;
+}
